+++ CHANGELOG.md
@@ -5,7 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
-No notable changes.
+### Added
+
+* Added `Rayne\Ecoji\EcojiStream` for stream encoding and decoding
 
 ## [1.0.1] - 2018-03-14